+++ packages/waggle/src/socket/DataServer.ts
@@ -3,11 +3,12 @@ import { createServer, Server } from 'http'
 const socketio = require('socket.io')
 
 export class DataServer {
-  public PORT: number = 4677
+  public PORT: number
   private readonly _app: express.Application
   private readonly server: Server
   public io: SocketIO.Server
-  constructor() {
+  constructor(port?: number) {
+    this.PORT = port || 4677
     this._app = express()
     this.server = createServer(this._app)
     this.initSocket()